requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
Flask==2.3.3
Flask-Cors==4.0.0
frozenlist==1.4.0
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonify==0.5
MarkupSafe==2.1.3
multidict==6.0.4
packaging==23.2
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.6
Werkzeug==2.3.7
yarl==1.9.2
zipp==3.17.0
